Refactor semantic highlighting with consolidated shared utilities
- Create unified utils package with consolidated logic
  - HighlightExtractorUtils: consolidates field text, model ID, query text extraction
  - HighlightValidator: unified validation for single and batch modes
  - HighlightConfigBuilder: merged configuration building with validation

- Remove redundant batch classes (~600 lines of duplicate code)
  - Delete HighlightConfigExtractor (replaced by HighlightConfigBuilder)
  - Delete HighlightRequestValidator and batch HighlightValidator
  - Consolidate validation logic into single utility

- Update existing components to use new shared utilities
  - SemanticHighlighter: use HighlightExtractorUtils for extraction
  - SemanticHighlighterEngine: delegate to utils with deprecation markers
  - SemanticHighlightingProcessor: use unified HighlightConfigBuilder
  - SemanticHighlightingFactory: use utils for semantic field detection

- Maintain backward compatibility
  - Single and batch highlighting paths work independently
  - Deprecated methods maintain API compatibility
  - All existing functionality preserved

- Fix unit test compilation and update test assertions
  - Update tests to use new unified utilities instead of deleted classes
  - Fix ambiguous assertEquals calls in neural query tests
  - Add proper mocking for search response validation

+++ src/test/java/org/opensearch/neuralsearch/query/NeuralQueryBuilderTests.java
@@ -1082,7 +1082,7 @@ public class NeuralQueryBuilderTests extends OpenSearchTestCase {
         KNNQueryBuilder returnedKNNQueryBuilder = (KNNQueryBuilder) result;
         assertEquals("Field name should match", FIELD_NAME, returnedKNNQueryBuilder.fieldName());
         assertArrayEquals("Vector should match", testVector, (float[]) returnedKNNQueryBuilder.vector(), 0.0f);
-        assertEquals("K should match", neuralQueryBuilder.k().intValue(), returnedKNNQueryBuilder.getK());
+        assertEquals("K should match", neuralQueryBuilder.k().intValue(), (int) returnedKNNQueryBuilder.getK());
 
     }
 
@@ -1146,7 +1146,7 @@ public class NeuralQueryBuilderTests extends OpenSearchTestCase {
             DELTA_FOR_FLOATS_ASSERTION
         );
         // For radial search, KNNQueryBuilder defaults K to 0 when not specified
-        assertEquals("K should be 0 for radial search when not specified", 0, returnedKNNQueryBuilder.getK());
+        assertEquals("K should be 0 for radial search when not specified", 0, (int) returnedKNNQueryBuilder.getK());
     }
 
     public void testPrepareTwoPhase_Query_whenRawTokensAvailable() {
